+++ src/__tests__/dispatch.test.ts
@@ -2,8 +2,8 @@ import * as Dispatch from "../dispatch";
 
 test("Function is memoized", () => {
   const dispatch: Dispatch.Dispatch<string> = () => "AString";
-  const actionMapper: Dispatch.ActionMapper<string, string> = () => "parentaction";
-  const mappedDispatch1 = Dispatch.mapDispatch(actionMapper, dispatch);
-  const mappedDispatch2 = Dispatch.mapDispatch(actionMapper, dispatch);
+  const actionMapper: (a1: string) => string = () => "parentaction";
+  const mappedDispatch1 = Dispatch.map(actionMapper, dispatch);
+  const mappedDispatch2 = Dispatch.map(actionMapper, dispatch);
   expect(mappedDispatch1).toBe(mappedDispatch2);
 });