+++ src/index.ts
@@ -1,16 +1,17 @@
 /* eslint-disable import/first, import/newline-after-import, import/order */
 
-// Run-time
-export * from "./program";
-export * from "./runtime";
-export * from "./dispatch";
+// Program
+import * as ProgramNs from "./program";
+export const Program = ProgramNs;
+export type Program<S, A, V> = ProgramNs.Program<S, A, V>;
+
+// Dispatch
+import * as DispatchNs from "./dispatch";
+export const Dispatch = DispatchNs;
+export type Dispatch<A> = DispatchNs.Dispatch<A>;
+
 // Effect manager
 export * from "./effect-manager";
-
-// Result
-import * as ResultNs from "./result";
-export const Result = ResultNs;
-export type Result<TError, TValue> = ResultNs.Result<TError, TValue>;
 
 // Cmd
 import * as CmdNs from "./cmd";
@@ -21,3 +22,8 @@ export type Cmd<A> = CmdNs.Cmd<A>;
 import * as SubNs from "./sub";
 export const Sub = SubNs;
 export type Sub<A> = SubNs.Sub<A>;
+
+// Result
+import * as ResultNs from "./result";
+export const Result = ResultNs;
+export type Result<TError, TValue> = ResultNs.Result<TError, TValue>;